+++ packages/website/app/scrollend-polyfill/page.tsx
@@ -1,0 +1,3 @@
+import Readme from "@af-utils/scrollend-polyfill/README.md";
+
+export default Readme;