+++ packages/website/components/AutoLink/index.tsx
@@ -4,11 +4,29 @@ import { ComponentPropsWithoutRef } from "react";
 type AutoLinkProps = ComponentPropsWithoutRef<typeof Link>;
 
 const AutoLink = ({ href, prefetch = true, ...props }: AutoLinkProps) => {
-    if (href && typeof href === "string" && /^https?:\//.test(href)) {
-        return <a href={href} target="_blank" rel="noopener" {...props} />;
+    let normalizedHref = href;
+
+    if (href && typeof href === "string") {
+        normalizedHref = href.replace(
+            process.env.NEXT_PUBLIC_ORIGIN as string,
+            ""
+        );
+
+        if (normalizedHref.startsWith("https://")) {
+            return (
+                <a
+                    href={normalizedHref}
+                    target="_blank"
+                    rel="noopener"
+                    {...props}
+                />
+            );
+        }
+    } else {
+        normalizedHref = href;
     }
 
-    return <Link href={href} prefetch={prefetch} {...props} />;
+    return <Link href={normalizedHref} prefetch={prefetch} {...props} />;
 };
 
 export default AutoLink;