+++ packages/website/components/NavLink/index.tsx
@@ -7,20 +7,22 @@ import { ComponentPropsWithoutRef } from "react";
 
 type NavLinkProps = ComponentPropsWithoutRef<typeof Link> & {
     activeClassName: string;
+    exact?: boolean;
 };
 
 const NavLink = ({
     href,
+    exact,
     className,
     activeClassName,
     prefetch = true,
     ...props
 }: NavLinkProps) => {
     const pathname = usePathname();
-    const isActive =
-        href === "/virtual"
-            ? pathname === href
-            : typeof href === "string" && pathname.startsWith(href);
+
+    const isActive = exact
+        ? pathname === href
+        : typeof href === "string" && pathname.startsWith(href);
 
     return (
         <Link