+++ packages/website/components/layouts/Documentation/index.tsx
@@ -1,36 +1,26 @@
 import { ReactNode } from "react";
 import { VscMenu, VscClose } from "react-icons/vsc";
-import Menu from "components/Menu";
+import Menu, { MenuItem } from "components/Menu";
 import ClientMenu from "components/ClientMenu";
 import Link from "next/link";
 import { cx } from "@emotion/css";
-import glob from "fast-glob";
 
-const renderTopLink = (className: string) => (
+const renderTopLink = (className: string, productName: string) => (
     <h1 className={cx("flex gap-2 font-medium", className)}>
         <Link href="/" prefetch={false} className="underline">
             af-utils
         </Link>
-        /<span className="text-slate-500">virtual</span>
+        /<span className="text-slate-500">{productName}</span>
     </h1>
 );
 
-const Virtual = ({ children }: { children: ReactNode }) => {
-    const map = glob
-        .sync("./components/examples/react-examples/**/code.tsx")
-        .reduce(
-            (result, path) => (
-                path
-                    .replace(/^.+react-examples\//, "")
-                    .replace(/\/code.+$/, "")
-                    .split("/")
-                    .reduce((acc, v) => (acc[v] ||= {}), result),
-                result
-            ),
-            {} as Record<string, any>
-        );
-
-    return (
+const getDocumentationLayout =
+    (
+        items: readonly MenuItem[] | MenuItem[],
+        prefix: string,
+        productName: string
+    ) =>
+    ({ children }: { children: ReactNode }) => (
         <div className="h-screen w-screen flex flex-col lg:flex-row">
             <details className="lg:hidden ds-menu z-10 relative">
                 <summary className="list-none max-w-none border-b flex h-[60px] items-center px-4 gap-8">
@@ -48,22 +38,22 @@ const Virtual = ({ children }: { children: ReactNode }) => {
                         aria-label="Close menu"
                         role="button"
                     />
-                    {renderTopLink("text-lg")}
+                    {renderTopLink("text-lg", productName)}
                 </summary>
                 <ClientMenu
-                    map={map}
-                    className="overflow-auto p-4 fixed inset-0 top-[61px] bg-white"
+                    items={items}
+                    prefix={prefix}
+                    className="overflow-auto p-34 fixed inset-0 top-[61px] bg-white"
                 />
             </details>
             <aside className="shadow-lg hidden lg:flex h-screen overflow-y-scroll flex-none flex-col min-w-[20em]">
                 <div className="sticky top-0 bg-white p-5 border-b">
-                    {renderTopLink("text-xl")}
+                    {renderTopLink("text-xl", productName)}
                 </div>
-                <Menu map={map} className="p-4" />
+                <Menu items={items} prefix={prefix} className="p-4" />
             </aside>
             <main className="h-full flex-1 overflow-auto p-4">{children}</main>
         </div>
     );
-};
 
-export default Virtual;
+export default getDocumentationLayout;