+++ packages/website/mdx-components.tsx
@@ -1,11 +1,20 @@
 import AutoLink from "components/AutoLink";
-import type { ReactNode } from "react";
+import { cx } from "@emotion/css";
+import type { ComponentPropsWithoutRef, ElementType } from "react";
 import type { MDXComponents } from "mdx/types";
+
+type WrapperProps<T extends ElementType = "div"> =
+    ComponentPropsWithoutRef<T> & { component?: T; className?: string };
 
 export const useMDXComponents = (components: MDXComponents) => ({
     ...components,
-    wrapper: ({ children }: { children: ReactNode }) => (
-        <div className="prose">{children}</div>
-    ),
+    wrapper: <T extends ElementType = "div">({
+        className,
+        component,
+        children
+    }: WrapperProps<T>) => {
+        const C = component || "div";
+        return <C className={cx("prose", className)}>{children}</C>;
+    },
     a: AutoLink
 });